refactor Effects processors
