+++ src/ImageSharp/Processing/OilPaintExtensions.cs
@@ -1,14 +1,13 @@
 ﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing.Processors.Effects;
 using SixLabors.Primitives;
 
 namespace SixLabors.ImageSharp.Processing
 {
     /// <summary>
-    /// Adds oil painting effect extensions to the <see cref="Image{TPixel}"/> type.
+    /// Adds oil painting effect extensions to the <see cref="Image"/> type.
     /// </summary>
     public static class OilPaintExtensions
     {
@@ -16,43 +15,36 @@ namespace SixLabors.ImageSharp.Processing
         /// Alters the colors of the image recreating an oil painting effect with levels and brushSize
         /// set to <value>10</value> and <value>15</value> respectively.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
         /// <param name="source">The image this method extends.</param>
         /// <returns>The <see cref="Image{TPixel}"/>.</returns>
-        public static IImageProcessingContext<TPixel> OilPaint<TPixel>(this IImageProcessingContext<TPixel> source)
-            where TPixel : struct, IPixel<TPixel>
-            => OilPaint(source, 10, 15);
+        public static IImageProcessingContext OilPaint(this IImageProcessingContext source) => OilPaint(source, 10, 15);
 
         /// <summary>
         /// Alters the colors of the image recreating an oil painting effect  with levels and brushSize
         /// set to <value>10</value> and <value>15</value> respectively.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
         /// <param name="source">The image this method extends.</param>
         /// <param name="rectangle">
         /// The <see cref="Rectangle"/> structure that specifies the portion of the image object to alter.
         /// </param>
         /// <returns>The <see cref="Image{TPixel}"/>.</returns>
-        public static IImageProcessingContext<TPixel> OilPaint<TPixel>(this IImageProcessingContext<TPixel> source, Rectangle rectangle)
-            where TPixel : struct, IPixel<TPixel>
-            => OilPaint(source, 10, 15, rectangle);
+        public static IImageProcessingContext OilPaint(this IImageProcessingContext source, Rectangle rectangle) =>
+            OilPaint(source, 10, 15, rectangle);
 
         /// <summary>
         /// Alters the colors of the image recreating an oil painting effect.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
         /// <param name="source">The image this method extends.</param>
         /// <param name="levels">The number of intensity levels. Higher values result in a broader range of color intensities forming part of the result image.</param>
         /// <param name="brushSize">The number of neighboring pixels used in calculating each individual pixel value.</param>
         /// <returns>The <see cref="Image{TPixel}"/>.</returns>
-        public static IImageProcessingContext<TPixel> OilPaint<TPixel>(this IImageProcessingContext<TPixel> source, int levels, int brushSize)
-            where TPixel : struct, IPixel<TPixel>
-            => source.ApplyProcessor(new OilPaintingProcessor<TPixel>(levels, brushSize));
+        public static IImageProcessingContext
+            OilPaint(this IImageProcessingContext source, int levels, int brushSize) =>
+            source.ApplyProcessor(new OilPaintingProcessor(levels, brushSize));
 
         /// <summary>
         /// Alters the colors of the image recreating an oil painting effect.
         /// </summary>
-        /// <typeparam name="TPixel">The pixel format.</typeparam>
         /// <param name="source">The image this method extends.</param>
         /// <param name="levels">The number of intensity levels. Higher values result in a broader range of color intensities forming part of the result image.</param>
         /// <param name="brushSize">The number of neighboring pixels used in calculating each individual pixel value.</param>
@@ -60,8 +52,11 @@ namespace SixLabors.ImageSharp.Processing
         /// The <see cref="Rectangle"/> structure that specifies the portion of the image object to alter.
         /// </param>
         /// <returns>The <see cref="Image{TPixel}"/>.</returns>
-        public static IImageProcessingContext<TPixel> OilPaint<TPixel>(this IImageProcessingContext<TPixel> source, int levels, int brushSize, Rectangle rectangle)
-            where TPixel : struct, IPixel<TPixel>
-            => source.ApplyProcessor(new OilPaintingProcessor<TPixel>(levels, brushSize), rectangle);
+        public static IImageProcessingContext OilPaint(
+            this IImageProcessingContext source,
+            int levels,
+            int brushSize,
+            Rectangle rectangle) =>
+            source.ApplyProcessor(new OilPaintingProcessor(levels, brushSize), rectangle);
     }
 }
