+++ tests/ImageSharp.Tests/Processing/Effects/OilPaintTest.cs
@@ -14,7 +14,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Effects
         public void OilPaint_OilPaintingProcessorDefaultsSet()
         {
             this.operations.OilPaint();
-            var processor = this.Verify<OilPaintingProcessor<Rgba32>>();
+            var processor = this.Verify<OilPaintingProcessor>();
 
             Assert.Equal(10, processor.Levels);
             Assert.Equal(15, processor.BrushSize);
@@ -24,7 +24,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Effects
         public void OilPaint_rect_OilPaintingProcessorDefaultsSet()
         {
             this.operations.OilPaint(this.rect);
-            var processor = this.Verify<OilPaintingProcessor<Rgba32>>(this.rect);
+            var processor = this.Verify<OilPaintingProcessor>(this.rect);
 
             Assert.Equal(10, processor.Levels);
             Assert.Equal(15, processor.BrushSize);
@@ -33,7 +33,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Effects
         public void OilPaint_Levels_Brsuh_OilPaintingProcessorDefaultsSet()
         {
             this.operations.OilPaint(34, 65);
-            var processor = this.Verify<OilPaintingProcessor<Rgba32>>();
+            var processor = this.Verify<OilPaintingProcessor>();
 
             Assert.Equal(34, processor.Levels);
             Assert.Equal(65, processor.BrushSize);
@@ -43,7 +43,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Effects
         public void OilPaint_Levels_Brsuh_rect_OilPaintingProcessorDefaultsSet()
         {
             this.operations.OilPaint(54, 43, this.rect);
-            var processor = this.Verify<OilPaintingProcessor<Rgba32>>(this.rect);
+            var processor = this.Verify<OilPaintingProcessor>(this.rect);
 
             Assert.Equal(54, processor.Levels);
             Assert.Equal(43, processor.BrushSize);
