+++ tests/ImageSharp.Tests/Processing/Effects/PixelateTest.cs
@@ -14,7 +14,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Effects
         public void Pixelate_PixelateProcessorDefaultsSet()
         {
             this.operations.Pixelate();
-            var processor = this.Verify<PixelateProcessor<Rgba32>>();
+            var processor = this.Verify<PixelateProcessor>();
 
             Assert.Equal(4, processor.Size);
         }
@@ -23,7 +23,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Effects
         public void Pixelate_Size_PixelateProcessorDefaultsSet()
         {
             this.operations.Pixelate(12);
-            var processor = this.Verify<PixelateProcessor<Rgba32>>();
+            var processor = this.Verify<PixelateProcessor>();
 
             Assert.Equal(12, processor.Size);
         }
@@ -32,7 +32,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Effects
         public void Pixelate_Size_rect_PixelateProcessorDefaultsSet()
         {
             this.operations.Pixelate(23, this.rect);
-            var processor = this.Verify<PixelateProcessor<Rgba32>>(this.rect);
+            var processor = this.Verify<PixelateProcessor>(this.rect);
 
             Assert.Equal(23, processor.Size);
         }
